Fix Data Modell

+++ CheckDiePreise/Data/Models/ProductChange.cs
@@ -12,9 +12,9 @@ namespace CheckDiePreise.Data.Models
         [Required] public string Name { get; set; }
         [Required] public DateTime Date { get; set; }
         [Required] public string Identifier { get; set; }
-        [Required] public float PriceUnit { get; set; }
+        [Required] public decimal PriceUnit { get; set; }
         [Required] public string UnitName { get; set; }
-        [Required] public float Baseprice { get; set; }
+        [Required] public decimal Baseprice { get; set; }
         [Required] public string BasepriceName { get; set; }
         [Required] public string Store { get; set; }
         public string Category { get; set; }
